+++ Database/MongoDB/Internal/Protocol.hs
@@ -8,6 +8,14 @@
 {-# LANGUAGE CPP, FlexibleContexts, TupleSections, TypeSynonymInstances #-}
 {-# LANGUAGE MultiParamTypeClasses, FlexibleInstances, UndecidableInstances #-}
 
+{-# LANGUAGE NamedFieldPuns, ScopedTypeVariables #-}
+
+#if (__GLASGOW_HASKELL__ >= 706)
+{-# LANGUAGE RecursiveDo #-}
+#else
+{-# LANGUAGE DoRec #-}
+#endif
+
 module Database.MongoDB.Internal.Protocol (
     FullCollection,
     -- * Pipe
@@ -19,13 +27,13 @@ module Database.MongoDB.Internal.Protocol (
     -- ** Reply
     Reply(..), ResponseFlag(..),
     -- * Authentication
-    Username, Password, Nonce, pwHash, pwKey
+    Username, Password, Nonce, pwHash, pwKey,
+    isClosed, close
 ) where
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative ((<$>))
 #endif
-import Control.Arrow ((***))
 import Control.Monad (forM, replicateM, unless)
 import Data.Binary.Get (Get, runGet)
 import Data.Binary.Put (Put, runPut)
@@ -35,6 +43,12 @@ import Data.IORef (IORef, newIORef, atomicModifyIORef)
 import System.IO (Handle)
 import System.IO.Unsafe (unsafePerformIO)
 import Data.Maybe (maybeToList)
+import GHC.Conc (ThreadStatus(..), threadStatus)
+import Control.Monad (forever)
+import Control.Concurrent.Chan (Chan, newChan, readChan, writeChan)
+import Control.Concurrent (ThreadId, forkIO, killThread)
+
+import Control.Exception.Lifted (onException, throwIO, try)
 
 import qualified Data.ByteString.Lazy as L
 
@@ -48,38 +62,111 @@ import qualified Crypto.Hash.MD5 as MD5
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as TE
 
-import Database.MongoDB.Internal.Util (whenJust, bitOr, byteStringHex)
-import System.IO.Pipeline (Pipeline, newPipeline, IOStream(..))
+import Database.MongoDB.Internal.Util (bitOr, byteStringHex)
 
-import qualified System.IO.Pipeline as P
+import Database.MongoDB.Transport (Transport)
+import qualified Database.MongoDB.Transport as T
 
-import Database.MongoDB.Internal.Connection (Connection)
-import qualified Database.MongoDB.Internal.Connection as Connection
+#if MIN_VERSION_base(4,6,0)
+import Control.Concurrent.MVar.Lifted (MVar, newEmptyMVar, newMVar, withMVar,
+                                       putMVar, readMVar, mkWeakMVar)
+#else
+import Control.Concurrent.MVar.Lifted (MVar, newEmptyMVar, newMVar, withMVar,
+                                         putMVar, readMVar, addMVarFinalizer)
+#endif
+
+#if !MIN_VERSION_base(4,6,0)
+mkWeakMVar :: MVar a -> IO () -> IO ()
+mkWeakMVar = addMVarFinalizer
+#endif
+
+-- * Pipeline
+
+-- | Thread-safe and pipelined connection
+data Pipeline = Pipeline {
+    vStream :: MVar Transport,  -- ^ Mutex on handle, so only one thread at a time can write to it
+    responseQueue :: Chan (MVar (Either IOError Response)),  -- ^ Queue of threads waiting for responses. Every time a response arrive we pop the next thread and give it the response.
+    listenThread :: ThreadId
+    }
+
+-- | Create new Pipeline over given handle. You should 'close' pipeline when finished, which will also close handle. If pipeline is not closed but eventually garbage collected, it will be closed along with handle.
+newPipeline :: Transport -> IO Pipeline
+newPipeline stream = do
+    vStream <- newMVar stream
+    responseQueue <- newChan
+    rec
+        let pipe = Pipeline{..}
+        listenThread <- forkIO (listen pipe)
+    _ <- mkWeakMVar vStream $ do
+        killThread listenThread
+        T.close stream
+    return pipe
+
+close :: Pipeline -> IO ()
+-- ^ Close pipe and underlying connection
+close Pipeline{..} = do
+    killThread listenThread
+    T.close =<< readMVar vStream
+
+isClosed :: Pipeline -> IO Bool
+isClosed Pipeline{listenThread} = do
+    status <- threadStatus listenThread
+    return $ case status of
+        ThreadRunning -> False
+        ThreadFinished -> True
+        ThreadBlocked _ -> False
+        ThreadDied -> True
+--isPipeClosed Pipeline{..} = isClosed =<< readMVar vHandle  -- isClosed hangs while listen loop is waiting on read
+
+listen :: Pipeline -> IO ()
+-- ^ Listen for responses and supply them to waiting threads in order
+listen Pipeline{..} = do
+    stream <- readMVar vStream
+    forever $ do
+        e <- try $ readMessage stream
+        var <- readChan responseQueue
+        putMVar var e
+        case e of
+            Left err -> T.close stream >> ioError err  -- close and stop looping
+            Right _ -> return ()
+
+psend :: Pipeline -> Message -> IO ()
+-- ^ Send message to destination; the destination must not response (otherwise future 'call's will get these responses instead of their own).
+-- Throw IOError and close pipeline if send fails
+psend p@Pipeline{..} message = withMVar vStream (flip writeMessage message) `onException` close p
+
+pcall :: Pipeline -> Message -> IO (IO Response)
+-- ^ Send message to destination and return /promise/ of response from one message only. The destination must reply to the message (otherwise promises will have the wrong responses in them).
+-- Throw IOError and closes pipeline if send fails, likewise for promised response.
+pcall p@Pipeline{..} message = withMVar vStream doCall `onException` close p  where
+    doCall stream = do
+        writeMessage stream message
+        var <- newEmptyMVar
+        liftIO $ writeChan responseQueue var
+        return $ readMVar var >>= either throwIO return -- return promise
 
 -- * Pipe
 
-type Pipe = Pipeline Response Message
+type Pipe = Pipeline
 -- ^ Thread-safe TCP connection with pipelined requests
 
 newPipe :: Handle -> IO Pipe
 -- ^ Create pipe over handle
-newPipe handle = Connection.fromHandle handle >>= newPipeWith
+newPipe handle = T.fromHandle handle >>= newPipeWith
 
-newPipeWith :: Connection -> IO Pipe
+newPipeWith :: Transport -> IO Pipe
 -- ^ Create pipe over connection
-newPipeWith conn = newPipeline $ IOStream (writeMessage conn)
-                                          (readMessage conn)
-                                          (Connection.close conn)
+newPipeWith conn = newPipeline conn
 
 send :: Pipe -> [Notice] -> IO ()
 -- ^ Send notices as a contiguous batch to server with no reply. Throw IOError if connection fails.
-send pipe notices = P.send pipe (notices, Nothing)
+send pipe notices = psend pipe (notices, Nothing)
 
 call :: Pipe -> [Notice] -> Request -> IO (IO Reply)
 -- ^ Send notices and request as a contiguous batch to server and return reply promise, which will block when invoked until reply arrives. This call and resulting promise will throw IOError if connection fails.
 call pipe notices request = do
     requestId <- genRequestId
-    promise <- P.call pipe (notices, Just (request, requestId))
+    promise <- pcall pipe (notices, Just (request, requestId))
     return $ check requestId <$> promise
  where
     check requestId (responseTo, reply) = if requestId == responseTo then reply else
@@ -91,7 +178,7 @@ type Message = ([Notice], Maybe (Request, RequestId))
 -- ^ A write notice(s) with getLastError request, or just query request.
 -- Note, that requestId will be out of order because request ids will be generated for notices after the request id supplied was generated. This is ok because the mongo server does not care about order just uniqueness.
 
-writeMessage :: Connection -> Message -> IO ()
+writeMessage :: Transport -> Message -> IO ()
 -- ^ Write message to connection
 writeMessage conn (notices, mRequest) = do
     noticeStrings <- forM notices $ \n -> do
@@ -104,8 +191,8 @@ writeMessage conn (notices, mRequest) = do
           let s = runPut $ putRequest request requestId
           return $ (lenBytes s) `L.append` s
 
-    Connection.write conn $ L.toStrict $ L.concat $ noticeStrings ++ (maybeToList requestString)
-    Connection.flush conn
+    T.write conn $ L.toStrict $ L.concat $ noticeStrings ++ (maybeToList requestString)
+    T.flush conn
  where
     lenBytes bytes = encodeSize . toEnum . fromEnum $ L.length bytes
     encodeSize = runPut . putInt32 . (+ 4)
@@ -113,12 +200,12 @@ writeMessage conn (notices, mRequest) = do
 type Response = (ResponseTo, Reply)
 -- ^ Message received from a Mongo server in response to a Request
 
-readMessage :: Connection -> IO Response
+readMessage :: Transport -> IO Response
 -- ^ read response from a connection
 readMessage conn = readResp  where
     readResp = do
-        len <- fromEnum . decodeSize <$> Connection.readExactly conn 4
-        runGet getReply <$> Connection.readExactly conn len
+        len <- fromEnum . decodeSize . L.fromStrict <$> T.read conn 4
+        runGet getReply . L.fromStrict <$> T.read conn len
     decodeSize = subtract 4 . runGet getInt32
 
 type FullCollection = Text