+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Package Versioning Policy](https://wiki.haskell.org/Package_versioning_policy).
 
+## [2.1.0] - unreleased
+
+### Added
+- TLS implementation. So far it is an experimental feature.
+
+### Removed
+- System.IO.Pipeline module
+
 ## [2.0.10] - 2015-12-22
 
 ### Fixed